% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Estimate the baseline for each m/z channel}
\usage{
baseline(macha, ppmwin = 3, lambda1 = 3, lambda2 = 6)
}
\arguments{
\item{macha}{Macha list containing list items k and s.}

\item{ppmwin}{Numeric. The ppm window within which to sum to denote the baseline.}

\item{lambda1, lambda2}{Integer. See \code{\link[baseline::baseline.irls]}}
}
\value{
List (a Macha object) containing additional list named k_b containing the calculated baseline intensity at every mass peak. (Note that this refers to peak IDs much like a relational database.)
}
\description{
\code{baseline} estimates the baseline noise level at each point supplied in \code{roi}
}
\details{
Developed to robustly estimate baselines in a variety of situations. Primarily: 1. cases where the baseline is not always observed, as in QE data; 2. Cases where a single ROI does not capture a large enough region to estimate the baseline.
To achieve this, ROIs of similar mass are grouped based on \code{ppmwin}. For each ROI group all the signal throughout the dataset is aggregated into an EIC. This EIC roll-joined, such that 0 values are filled with the last observed intensity. From this EIC a baseline is estimated.
}
\examples{
\dontrun{
baseline(macha, ppmwin = 3, lambda1 = 6, lambda2 = 7)
}

}
\seealso{
\code{\link[baseline::baseline.irls]}
}
