% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findrois.R
\name{centroidgroups}
\alias{centroidgroups}
\title{Speed minded function for iteratively splitting a group of points by two density estimates.}
\usage{
centroidgroups(features, ppm, rtwid, minlength)
}
\arguments{
\item{features}{Matrix. Rows are observed mass peaks. Columns are ordered to contain "mz", "rt", "i".}

\item{ppm}{Numeric. Bandwidth of kernel for estimating density in the mz dimension.}

\item{rtwid}{Numeric. Bandwidth of kernel for estimating density in the rt dimension.}

\item{minlength}{Integer. If an ROI is shorter than this it is discarded.}
}
\value{
Integer vector containing assignment of each point to a group.
}
\description{
\code{centroidgroups} takes mass spectra and assigns peaks to ROIs.
}
