% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcomponents.R
\name{findcomponents}
\alias{findcomponents}
\title{Analyze ROIs for retention-time-domain peaks satisfying parameters}
\usage{
findcomponents(macha, unrelated.dist = 40, min.peakwidth = 3,
  min.sharpness = 2, min.fracobs = 0.4)
}
\arguments{
\item{unrelated.dist}{Integer. The distance two seeds must be apart to fit them individually (much faster). Suggested value: large peakwidth * 3}

\item{min.peakwidth}{numeric. The minimum width a component can be to be retained. In seconds.}

\item{min.sharpness}{numeric. A lower limit on peak.height/peak.width.  Units Intensity/Second. (Not very diagnostic.)}

\item{min.fractionobs}{numeric. A lower limit on the fraction of expected observations present in the peak region.}
}
\value{
List (a Macha object) containing additional list named $c containing detected components and the parameters of the skew normal distribution describing that component. (and $c_error)
}
\description{
\code{findcomponents} Sequentially analyzes detected ROIs for peaks. This proceeds by several steps.
}
\details{
1. Peaks are seeded by a fast, scale-space peak detection algorithm
2. Skew-normal distributions are initialized and quickly fit to the seed regions.3
3. Distributions within \code{unrelated.dist} of each other are fit simultaneously to the raw data.
4. Fit distributions are checked by the supplied parameters and peaks not satisfying those parameters are removed.  The distribution is then refit.
5. This repeats until the fit distributions stabilize.
}
