% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findrois.R
\name{findrois}
\alias{findrois}
\title{ROI detection based on density estimation.}
\usage{
findrois(features, minlength = 5, ppm = 1, rtwid = 10)
}
\value{
List (a Macha object) containing additional list named k_r containing the assignment of each mass peak (k) to an ROI (r).
}
\description{
\code{findrois} takes mass spectra and assigns peaks to ROIs
}
\details{
In addition this function assures that only one point per scan is in a ROI.

See: \code{\link{centroidgroups}} for parameter descriptions.
}
\examples{
\dontrun{
findrois(macha, minlength = 15, ppm = 2, rtwid = 7)
}

}
