% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitseeds.R
\name{fitandreseed}
\alias{fitandreseed}
\title{Find a set of parameterized curves which describe the ROI well}
\usage{
fitandreseed(roi, seeds, unrelated.dist = 30, min.peakwidth = 3,
  sn.adjust = 1, min.sharpness = 1, min.fracobs = 0.4, do.plot = F)
}
\arguments{
\item{roi}{data.table. ROI}

\item{seeds}{matrix. Columns "start", "end", "peak". Rows are individual initialization points for ROI fitting. Values correspond to the index within the supplied ROI.}

\item{unrelated.dist}{Integer. The distance two seeds must be to fit them individually (much faster).}

\item{min.peakwidth}{numeric. The minimum width a component can be to be retained. In seconds.}

\item{sn.adjust}{numeric. A multiplier for the signal to noise limit of retained components. Calculated from the raw data and the supplied baseline.  Used to determine the number of observations as well.}

\item{min.sharpness}{numeric. A lower limit on peak.height/peak.width.  Units Intensity/Second.}

\item{do.plot}{boolean. Plot a bunch of random stuff.}

\item{min.fractionobs}{numeric. A lower limit on the fraction of expected observations present in the peak region.}
}
\value{
List of matrices. Matrices include the parameterization for each component. Eg. location, dispersion, skew, scaling factor, baseline. Additional 95th percentile values are also returned as rtmin and rtmax.
}
\description{
\code{fitandreseed} Iteritively fits a sum of Skew-Normal distibutions to the ROI until quality parameters are met.
}
\details{
Used to provide a parameterized rrepresentation of the ROI
}
