% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_grtgmzcor.R
\name{grtgmzcor}
\alias{grtgmzcor}
\title{Use well behaved peak groups to correct drift in both retention time and mass across samples}
\usage{
grtgmzcor(Nmacha, shaperng = 2, fracobs = 0.6)
}
\arguments{
\item{Nmacha}{An Nmacha object containing peaks grouped between samples.}

\item{shaperng}{Peaks within a group must exhibit a range of shape parameters less than this value to be included in the fit.}

\item{fracobs}{Peak groups must contain at least this fraction of peaks to be included in the fit.}
}
\value{
List (an Nmacha object) containing additional list named grt (mapping between grt and rt for each file), gmz (mapping between gmz and mz for each file), gc.
}
\description{
\code{grtmzcor} Determines global correction curves (one correction per scan) for both mz and rt and builds a mapping between the global retention time (rt.g) and observed retention time (rt)
}
\details{
A loess curve is fit to the difference between the observed value and the mean of all values against the retention time.
}
\examples{
\dontrun{
grtgmzcor(Nmacha, shaperng = 2, fracobs = .4)
}

}
