% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makemchans.R
\name{makemchans}
\alias{makemchans}
\title{Reconstruct ROIs split due to missing observations}
\usage{
makemchans(macha, ppmwin = 3, rtwin = 30)
}
\arguments{
\item{macha}{macha. Three columns: mz, s, k and (optional) g. s is scan, k is a unique integer ID for each peak, g is the initial group assignment.}

\item{ppmwin}{numeric. Mass distance in ppm to disqualify ROI aggregation. Recommended: instrument_ppm}

\item{rtwin}{numeric. RT distance in seconds to disqualify ROI aggregation. Recommended: 1/5 of run length}
}
\value{
macha. macha object with additional item $r_mchan
}
\description{
\code{makemchans} Find's ROIs which are likely the same mass channel.
}
\details{
Combining ROIs which were previously split due to missing values or conflicts in the webtrace step allows more complete information to be supplied to the baselining step.
}
