% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_roi.R
\name{makeroicache}
\alias{makeroicache}
\title{Compile ROIs}
\usage{
makeroicache(macha)
}
\arguments{
\item{macha}{macha.}
}
\value{
data.table
}
\description{
\code{makeroicache} Compiles ROIs rolling missing values and aggregating baselines and other calculated information.
}
