% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webroi.R
\name{rectwebtrace}
\alias{rectwebtrace}
\title{Mass trace detection first applying rectroi and then webroi to resolve conflicts}
\usage{
rectwebtrace(k, ppm.rect = 8, scan.rect = 4, scan.web = 20,
  neighbors.web = 3)
}
\arguments{
\item{ppm.rect}{numeric.}

\item{scan.rect}{integer.}

\item{scan.web}{integer.}

\item{neighbors.web}{integer.}
}
\value{
Integer vector. Each integer is a group assignment. In the same order as supplied k.
}
\description{
\code{rectwebtrace} Splits peaks based on mass and scan spacing.
}
\details{
Functional as a slower, robust ROI detection sensitive to shifts in mass. Necessary for high resolution and high-variance mass data.

See other functions for parameter definitions.
}
\seealso{
\code{\link[rectroi]} \code{\link[webtrace]}
}
