% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedpeaks.R
\name{seedpeaks}
\alias{seedpeaks}
\title{Find initial guesses for peaks and valleys}
\usage{
seedpeaks(roi, S = 3:8, seed.maxdensity = 1/5, seed.maxdist = 4,
  seed.sn.perpeak = c(Inf, 50, 5, 3, 2), seed.sn.range = 2,
  seed.minwidth = 3, do.plot = F)
}
\arguments{
\item{roi}{Matrix. Rows are observed mass peaks. Columns named "mz", "rt", "s", "b", "i".  Column "b" is the baseline at that point.}

\item{S}{Integer vector. The smoothing scales to apply when searching for peaks.}

\item{seed.maxdensity}{Float. Maximum density in which to seed peaks in peaks/scan. Eg 1/5.}

\item{seed.maxdist}{Integer. The maximum distance in scans the local maximum of the smoothed trace can travel and still positively reinforce a peak.}

\item{seed.sn.perpeak}{Numeric vector. Indices of the vector correspond to the number of observations which must be made of the mass in the peak region defined by seed.sn.range. The value at each index defines the minimum intensity necessary for an observation to count in value * baseline. Eg. c(Inf, 10) means a seed with only 1 observation will not be returned whereas a seed with 2 observations at 10*baseline will pass.}

\item{seed.sn.range}{Integer. Range in scans around a peak loci in which to look for observations qualifying as in seed.sn.perpeak.}

\item{seed.minwidth}{Numeric. Minimum width in scans of a putative region to be returned as a seed.}

\item{do.plot}{Boolean. Plot a bunch of random things.}
}
\value{
Matrix. Columns "start", "end", "peak", correspond to the indices of the peak and its flanking valleys.
}
\description{
\code{seedpeaks} Uses "scale space peak picking" to quickly find robust local maxima and minima
}
\details{
Used to provide an initialization for further peak modeling and component recovery.
}
\seealso{
\code{\link{sspp}}
}
