% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webroi.R
\name{webtrace}
\alias{webtrace}
\title{Split up peaks based on weighted graph of nearby mass observations}
\usage{
webtrace(k, scan, neighbors = 3)
}
\arguments{
\item{k}{data.table. Three columns: mz, s, k, g. s is scan, k is a unique integer ID for each peak, g is the initial group assignment.}

\item{scan}{integer. Scan window within to look for similar mass peaks. Suggested: 10.}

\item{neighbors}{integer. Within each window, for each mass, how many nearest neighbors to link. Suggested: 3}
}
\value{
Integer vector. Each integer is a group assignment. In the same order as supplied k.
}
\description{
\code{webtrace} Performs refinement of high resolution ROIs correcting errors due to nearby peaks, peaks coelescence, and multiple peaks per scan.
}
\details{
Splits peaks based on mass and scan spacing. A graph is formed by linking masses of nearest m/z values within a scan window.  The mincut of the graph which segregates conflicting ROIs (two mass peaks in a single scan) is determined.  The graph is iteratively cut until no conflicts remain.

Functional as a slower, robust ROI detection sensitive to shifts in mass. Necessary for high resolution and high-variance mass data.
}
\seealso{
\code{\link[rectroi]}
}
