% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedpeaks.R
\name{seedpeaks}
\alias{seedpeaks}
\title{Find initial guesses for peaks and valleys}
\usage{
seedpeaks(roi, S = 3:8, seed.maxdensity = 1/5, seed.maxdist = 4,
  seed.sn.perpeak = c(Inf, 50, 5, 3, 2), seed.sn.range = 2,
  seed.sn.adjust = 1, seed.minwidth = 3, do.plot = F)
}
\arguments{
\item{roi}{Matrix. Rows are observed mass peaks. Columns named "mz", "rt", "s", "b", "i".  Column "b" is the baseline at that point.}

\item{S}{Integer vector. The smoothing scales to apply when searching for peaks.}

\item{seed.maxdist}{Integer. The maximum distance in scans the local maximum of the smoothed trace can travel and still positively reinforce a peak.}

\item{seed.sn.adjust}{Numeric. Multiplied by the signal to noise limit.  The limit is variable but a value of 1 here corresponds to a SN range of 1.5 to 2.5.  See below.}

\item{seed.minwidth}{Numeric. Minimum width in scans of a putative region to be returned as a seed.}

\item{do.plot}{Boolean. Plot a bunch of random things.}

\item{noisy.chan.n}{Integer. The number of scans a mass-channel must be observed in to be assigned the maximum SN limit.  Defaults to 0.8*total.scans}
}
\value{
Matrix. Columns "start", "end", "peak", correspond to the indices of the peak and its flanking valleys.
}
\description{
\code{seedpeaks} Uses "scale space peak picking" to quickly find robust local maxima and minima
}
\details{
Used to provide an initialization for further peak modeling and component recovery.
}
\seealso{
\code{\link[sspp]}
}
